+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ## 1.54.1 (Unreleased)
+
+ENHANCEMENTS:
+
+* Resource `tencentcloud_tcr_vpc_attachment` add `enable_public_domain_dns`, `enable_vpc_domain_dns` to set whether to enable dns.
+* Data Source `tencentcloud_tcr_vpc_attachments` add `enable_public_domain_dns`, `enable_vpc_domain_dns`.
+
 ## 1.54.0 (March 22, 2021)
 
 FEATURES: